+++ packages/zwave-js/src/lib/node/mixins/05_Security.ts
@@ -12,7 +12,7 @@ import { getEnumMemberName } from "@zwave-js/shared";
 import { type Driver } from "../../driver/Driver.js";
 import { cacheKeys } from "../../driver/NetworkCache.js";
 import { type DeviceClass } from "../DeviceClass.js";
-import { NetworkRoleMixin } from "./001_NetworkRole.js";
+import { NetworkRoleMixin } from "./01_NetworkRole.js";
 
 export abstract class NodeSecurityMixin extends NetworkRoleMixin
 	implements QuerySecurityClasses, SetSecurityClass
