+++ packages/zwave-js/src/lib/node/mixins/30_Wakeup.ts
@@ -5,7 +5,7 @@ import {
 	ZWaveError,
 	ZWaveErrorCodes,
 } from "@zwave-js/core";
-import { NodeStatusMixin } from "./020_Status.js";
+import { NodeStatusMixin } from "./20_Status.js";
 
 /**
  * Interface for NodeWakeupMixin
