+++ packages/zwave-js/src/lib/node/mixins/40_Values.ts
@@ -20,7 +20,7 @@ import { type Driver } from "../../driver/Driver.js";
 import { type DeviceClass } from "../DeviceClass.js";
 import { type ZWaveNodeValueEventCallbacks } from "../_Types.js";
 import * as nodeUtils from "../utils.js";
-import { NodeWakeupMixin } from "./030_Wakeup.js";
+import { NodeWakeupMixin } from "./30_Wakeup.js";
 
 /** Defines functionality of Z-Wave nodes related to the value DB */
 export interface NodeValues {
