+++ packages/zwave-js/src/lib/node/mixins/80_DeviceConfig.ts
@@ -2,7 +2,7 @@ import { type DeviceConfig } from "@zwave-js/config";
 import { InterviewStage, type MaybeNotKnown, NOT_KNOWN } from "@zwave-js/core";
 import { Bytes, formatId } from "@zwave-js/shared";
 import { cacheKeys } from "../../driver/NetworkCache.js";
-import { FirmwareUpdateMixin } from "./070_FirmwareUpdate.js";
+import { FirmwareUpdateMixin } from "./70_FirmwareUpdate.js";
 
 export interface NodeDeviceConfig {
 	/**
