+++ packages/zwave-js/src/lib/node/mixins/index.ts
@@ -1,3 +1,3 @@
-import { DeviceConfigMixin } from "./080_DeviceConfig.js";
+import { DeviceConfigMixin } from "./80_DeviceConfig.js";
 
 export abstract class ZWaveNodeMixins extends DeviceConfigMixin {}
